web/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask-appbuilder>=4.8.0",
    "python-dotenv>=1.1.1",
    "flask-cors>=5.0.0",
    "opencv-python>=4.8.0",
    "mediapipe>=0.10.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
]
